Serperate directories for client and server
