+++ client/difficult.txt
@@ -1,0 +1,1 @@
+Dang this assignment is difficult for our first time doing this type of stuff. A lot of things to learn and get working... I hope we can figure this out.
