--- ls.txt
@@ -1,14 +1,0 @@
-1
-ftp_client.c
-ftp_client.exe
-ftp_client.exe.stackdump
-ftp_client_h.c
-ftp_server.c
-ftp_server.exe
-ftp_server_h.c
-hello.txt
-ls.txt
-stuff.c
-stuff2.c
-udp_client.c
-udp_server.c
