+++ server/random.txt
@@ -1,0 +1,1 @@
+lorem ipsum
